+++ src/services/investments.py
@@ -7,16 +7,17 @@ from ..models import (
     calculate_future_value as _calculate_future_value,
     get_total_projected_wealth as _get_total_projected_wealth,
 )
+from decimal import Decimal
 
-def add_or_update(name: str, current_value: float, monthly: float = 0, return_rate: float = 10.0, target_year: int = 2050, notes: str = ""):
+def add_or_update(name: str, current_value: Decimal, monthly: Decimal = 0, return_rate: Decimal = 10.0, target_year: int = 2050, notes: str = ""):
     add_or_update_investment(name=name, current_value=current_value, monthly=monthly, return_rate=return_rate, target_year=target_year, notes=notes)
 
 def get_investments() -> List[Investment]:
     return _get_investment()
 
-def calculate_future_value_for(inv: Investment, extra_monthly: float = 0) -> float:
+def calculate_future_value_for(inv: Investment, extra_monthly: Decimal = 0) -> Decimal:
     return _calculate_future_value(inv, extra_monthly)
 
-def get_total_projected_wealth(target_year: int = None) -> float:
+def get_total_projected_wealth(target_year: int = None) -> Decimal:
     """Service wrapper — can add logging, caching, etc. later"""
     return _get_total_projected_wealth(target_year)
